Split backend specific serialization tests out into their own files.
